--- app/src/main/java/com/truefan/sample/app/data/api/ApiHelper.kt
@@ -1,6 +1,0 @@
-package com.truefan.sample.app.data.api
-
-class ApiHelper(private val apiService: ApiService) {
-
-    suspend fun getData() = apiService.getData()
-}
