--- app/src/main/java/com/truefan/sample/app/data/api/ApiService.kt
@@ -1,11 +1,0 @@
-package com.truefan.sample.app.data.api
-
-import com.truefan.sample.app.data.model.Response
-import retrofit2.http.GET
-
-interface ApiService {
-
-    @GET("response")
-    suspend fun getData(): Response
-
-}
