--- app/src/main/java/com/truefan/sample/app/data/model/Response.kt
@@ -1,8 +1,0 @@
-package com.truefan.sample.app.data.model
-
-data class Response(
-    val quizFound : Boolean,
-    val rank : Int,
-    val score : Int,
-    val msg : String
-)
