--- app/src/main/java/com/truefan/sample/app/data/repository/MainRepository.kt
@@ -1,8 +1,0 @@
-package com.truefan.sample.app.data.repository
-
-import com.truefan.sample.app.data.api.ApiHelper
-
-class MainRepository(private val apiHelper: ApiHelper) {
-
-    suspend fun getData() = apiHelper.getData()
-}
