--- app/src/main/java/com/truefan/sample/app/ui/base/ViewModelFactory.kt
@@ -1,19 +1,0 @@
-package com.truefan.sample.app.ui.base
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.truefan.sample.app.data.api.ApiHelper
-import com.truefan.sample.app.data.repository.MainRepository
-import com.truefan.sample.app.ui.main.viewmodel.MainViewModel
-
-class ViewModelFactory(private val apiHelper: ApiHelper) : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
-            return MainViewModel(MainRepository(apiHelper)) as T
-        }
-        throw IllegalArgumentException("Unknown class name")
-    }
-
-}
-
