--- app/src/main/java/com/truefan/sample/app/ui/main/viewmodel/MainViewModel.kt
@@ -1,20 +1,0 @@
-package com.truefan.sample.app.ui.main.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.liveData
-import com.truefan.sample.app.data.repository.MainRepository
-import com.truefan.sample.app.utils.Resource
-import kotlinx.coroutines.Dispatchers
-
-
-class MainViewModel(private val mainRepository: MainRepository) : ViewModel() {
-
-    fun getData() = liveData(Dispatchers.IO) {
-        emit(Resource.loading(data = null))
-        try {
-            emit(Resource.success(data = mainRepository.getData()))
-        } catch (exception: Exception) {
-            emit(Resource.error(data = null, message = exception.message ?: "Error Occurred!"))
-        }
-    }
-}
