--- app/src/main/java/com/truefan/sample/app/utils/KotlinHelper.kt
@@ -1,13 +1,0 @@
-package com.truefan.sample.app.utils
-
-import android.view.LayoutInflater
-import androidx.activity.ComponentActivity
-import androidx.viewbinding.ViewBinding
-
-class KotlinHelper {
-    inline fun <T : ViewBinding> ComponentActivity.viewBinding(
-        crossinline bindingInflater: (LayoutInflater) -> T
-    ) = lazy(LazyThreadSafetyMode.NONE) {
-        bindingInflater.invoke(layoutInflater)
-    }
-}
