--- app/src/main/java/com/truefan/sample/app/utils/Resource.kt
@@ -1,17 +1,0 @@
-package com.truefan.sample.app.utils
-
-import com.truefan.sample.app.utils.Status.ERROR
-import com.truefan.sample.app.utils.Status.LOADING
-import com.truefan.sample.app.utils.Status.SUCCESS
-
-
-data class Resource<out T>(val status: Status, val data: T?, val message: String?) {
-    companion object {
-        fun <T> success(data: T): Resource<T> = Resource(status = SUCCESS, data = data, message = null)
-
-        fun <T> error(data: T?, message: String): Resource<T> =
-            Resource(status = ERROR, data = data, message = message)
-
-        fun <T> loading(data: T?): Resource<T> = Resource(status = LOADING, data = data, message = null)
-    }
-}
