--- app/src/main/java/com/truefan/sample/app/utils/Status.kt
@@ -1,7 +1,0 @@
-package com.truefan.sample.app.utils
-
-enum class Status {
-    SUCCESS,
-    ERROR,
-    LOADING
-}
