--- app/src/main/res/values/strings.xml
@@ -1,6 +1,0 @@
-<resources>
-    <string name="app_name">TrueFan Sample App</string>
-    <string name="rank">"Rank: "</string>
-    <string name="score">"Score: "</string>
-    <string name="msg">"Message: "</string>
-</resources>
