+++ build.gradle
@@ -2,6 +2,7 @@
 
 buildscript {
     ext.kotlin_version = '1.4.20'
+    ext.hilt_version = '2.38.1'
     repositories {
         google()
         jcenter()
@@ -9,7 +10,8 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:7.0.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.30"
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
