+++ retrofit_coroutine_mvvm_sample/proguard-rules.pro
