+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/data/api/ApiHelper.kt
@@ -1,0 +1,6 @@
+package com.retrofit.coroutine.mvvm.data.api
+
+class ApiHelper(private val apiService: ApiService) {
+
+    suspend fun getData() = apiService.getData()
+}
