+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/data/api/ApiService.kt
@@ -1,0 +1,11 @@
+package com.retrofit.coroutine.mvvm.data.api
+
+import com.retrofit.coroutine.mvvm.data.model.Response
+import retrofit2.http.GET
+
+interface ApiService {
+
+    @GET("response")
+    suspend fun getData(): Response
+
+}
