+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/data/api/RetrofitBuilder.kt
@@ -1,0 +1,17 @@
+package com.retrofit.coroutine.mvvm.data.api
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitBuilder {
+    private const val BASE_URL = "https://www.mocky.io/v2/5ec3eca9300000bed639c56e/"
+
+    private fun getRetrofit(): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+
+    val apiService: ApiService = getRetrofit().create(ApiService::class.java)
+}
