+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/data/model/Response.kt
@@ -1,0 +1,8 @@
+package com.retrofit.coroutine.mvvm.data.model
+
+data class Response(
+    val quizFound : Boolean,
+    val rank : Int,
+    val score : Int,
+    val msg : String
+)
