+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/data/repository/MainRepository.kt
@@ -1,0 +1,8 @@
+package com.retrofit.coroutine.mvvm.data.repository
+
+import com.retrofit.coroutine.mvvm.data.api.ApiHelper
+
+class MainRepository(private val apiHelper: ApiHelper) {
+
+    suspend fun getData() = apiHelper.getData()
+}
