+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/ui/base/ViewModelFactory.kt
@@ -1,0 +1,19 @@
+package com.retrofit.coroutine.mvvm.ui.base
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.retrofit.coroutine.mvvm.data.api.ApiHelper
+import com.retrofit.coroutine.mvvm.data.repository.MainRepository
+import com.retrofit.coroutine.mvvm.ui.main.viewmodel.MainViewModel
+
+class ViewModelFactory(private val apiHelper: ApiHelper) : ViewModelProvider.Factory {
+
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
+            return MainViewModel(MainRepository(apiHelper)) as T
+        }
+        throw IllegalArgumentException("Unknown class name")
+    }
+
+}
+
