+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/ui/main/viewmodel/MainViewModel.kt
@@ -1,0 +1,20 @@
+package com.retrofit.coroutine.mvvm.ui.main.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.liveData
+import com.retrofit.coroutine.mvvm.data.repository.MainRepository
+import com.retrofit.coroutine.mvvm.utils.Resource
+import kotlinx.coroutines.Dispatchers
+
+
+class MainViewModel(private val mainRepository: MainRepository) : ViewModel() {
+
+    fun getData() = liveData(Dispatchers.IO) {
+        emit(Resource.loading(data = null))
+        try {
+            emit(Resource.success(data = mainRepository.getData()))
+        } catch (exception: Exception) {
+            emit(Resource.error(data = null, message = exception.message ?: "Error Occurred!"))
+        }
+    }
+}
