+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/utils/KotlinHelper.kt
@@ -1,0 +1,13 @@
+package com.retrofit.coroutine.mvvm.utils
+
+import android.view.LayoutInflater
+import androidx.activity.ComponentActivity
+import androidx.viewbinding.ViewBinding
+
+class KotlinHelper {
+    inline fun <T : ViewBinding> ComponentActivity.viewBinding(
+        crossinline bindingInflater: (LayoutInflater) -> T
+    ) = lazy(LazyThreadSafetyMode.NONE) {
+        bindingInflater.invoke(layoutInflater)
+    }
+}
