+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/utils/Resource.kt
@@ -1,0 +1,17 @@
+package com.retrofit.coroutine.mvvm.utils
+
+import com.retrofit.coroutine.mvvm.utils.Status.ERROR
+import com.retrofit.coroutine.mvvm.utils.Status.LOADING
+import com.retrofit.coroutine.mvvm.utils.Status.SUCCESS
+
+
+data class Resource<out T>(val status: Status, val data: T?, val message: String?) {
+    companion object {
+        fun <T> success(data: T): Resource<T> = Resource(status = SUCCESS, data = data, message = null)
+
+        fun <T> error(data: T?, message: String): Resource<T> =
+            Resource(status = ERROR, data = data, message = message)
+
+        fun <T> loading(data: T?): Resource<T> = Resource(status = LOADING, data = data, message = null)
+    }
+}
