+++ retrofit_coroutine_mvvm_sample/src/main/java/com/retrofit/coroutine/mvvm/utils/Status.kt
@@ -1,0 +1,7 @@
+package com.retrofit.coroutine.mvvm.utils
+
+enum class Status {
+    SUCCESS,
+    ERROR,
+    LOADING
+}
