+++ retrofit_coroutine_mvvm_sample/src/main/res/drawable-v24/ic_launcher_foreground.xml
