+++ retrofit_coroutine_mvvm_sample/src/main/res/drawable/ic_launcher_background.xml
