+++ retrofit_coroutine_mvvm_sample/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
