+++ retrofit_coroutine_mvvm_sample/src/main/res/mipmap-hdpi/ic_launcher_round.png
