+++ retrofit_coroutine_mvvm_sample/src/main/res/mipmap-mdpi/ic_launcher.png
