+++ retrofit_coroutine_mvvm_sample/src/main/res/mipmap-xhdpi/ic_launcher_round.png
