+++ retrofit_coroutine_mvvm_sample/src/main/res/mipmap-xxhdpi/ic_launcher.png
