+++ retrofit_coroutine_mvvm_sample/src/main/res/mipmap-xxxhdpi/ic_launcher.png
