+++ retrofit_coroutine_mvvm_sample/src/main/res/values/colors.xml
