+++ retrofit_coroutine_mvvm_sample/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <string name="app_name">Retrofit Coroutine MVVM Sample</string>
+    <string name="rank">"Rank: "</string>
+    <string name="score">"Score: "</string>
+    <string name="msg">"Message: "</string>
+</resources>
