+++ retrofit_coroutine_mvvm_sample/src/test/java/com/retrofit/coroutine/mvvm/ExampleUnitTest.kt
@@ -1,0 +1,18 @@
+package com.retrofit.coroutine.mvvm
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
